Move releases.py to netbox/
