+++ netbox/netbox/tests/test_releases.py
